Amendment to RFQ
The United States Tax Court issued Request for Quote (RFQ) # 18-0018 on August 2, 2018.  The attached document is an amendment to that RFQ.  It contains supplemental information and Q&As 1, 8, 15, and 16 revise certain specific statements in the original RFQ.  Quotes are due by 5:00 p.m. EDT, Friday, August 24, 2018, and must be submitted by email to clerk@ustaxcourt.gov.